2 MesaWater Frontend changes aacording these tasks:
- The styling needs to be fixed, all normal sized buttons should be the same
 size. All large sized buttons should be the same size. 
The large buttons should be larger than the normal buttons.
- The text color needs to be changed to be ADA compliant.
- The secondary button should be the light blue color in the site logo.

 - Can we make the default padding between elements smaller so we dont have 
to add the additional style every time? Example above the space between the
 text w/menu section and basic section is too large.

- The main nav sub items need styling. Screenshot added. 

- Links within the content field in this section - image with icon and link
 - should stay white and have an underline. Here is an example , 
the linked phone number in the banner isnt ada compliant because of the color

- The content within the
 boxes gets cut off

- Increase space between lists and headings
could you please give example page?

Every time i add an image to the Text w/Image Fade section it is distorted

- The additional style Background: Light Blue (17) adds half of a wave? See
screenshot. I expected it to make the entire background of that sectio
n light blue.

- Styling for Financial services page